linuxKernel: keep track of kernelPatches passed to callPackage at call site
This is needed, so the boot.kernelPatches option in nixos can inject
additional patches properly. Otherwise it only has access to the
kernelPatches of the final derivation, which might or might not be what
was passed to the derivation in the original call.
The previous behaviour caused various hacks to become neccessary to
avoid duplicates in kernelPatches.